Decouple `activerecord.Conn`
This patch decouples `activerecord.Conn` interface into:
* `TransactionStatements`
* `DatabaseStatements`
* `SchemaStatements`

To ease implementation of the connection adapters.
